requirements copy.txt
Django==3.2.13
django-bootstrap4==21.1
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-fontawesome-5==1.0.18
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==1.4.6
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
whitenoise==5.3.0